# Optimized growth and array copy