@Salakar 2022-07-15
This is such a well put together PR with detailed benchmark which made it extremely easy to review and test out, thank you! I'm happy to ship this