@Salakar 2022-07-15
I'd also be happy to accept the growth benchmark as a PR, it'd be good to have to make sure no regressions.